requirements.txt
matplotlib
nltk
numpy
Pillow
argparse
pickle
pytorch
pandas
torchvision
